refactor(Scrap, Home, Calendar): 스크랩 관련 로직을 Home 서비스와 Calendar 서비스로 분리 및 DTO 계층 개선
